+++ src/server/render/render-tree-to-html.js
@@ -16,8 +16,11 @@ export default async ({
       queryParams
     }
   }
+  const data = Array.isArray(componentData)
+    ? { data: componentData }
+    : componentData
   // create html string from target component
-  const app = <Component {...componentData} _tapestry={_tapestryData} />
+  const app = <Component {...data} _tapestry={_tapestryData} />
   // getLoadableState must be called before renderToString to preload all import() components
   const loadableState = await getLoadableState(app)
   const htmlString = renderToString(app)
@@ -34,7 +37,7 @@ export default async ({
   const renderData = {
     ...styleData,
     head: helmet,
-    bootstrapData: componentData,
+    bootstrapData: data,
     _tapestryData,
     loadableState
   }